Mudando tipo das variaveis

+++ app/src/main/java/com/example/calculadorasimples/Calcular.kt
@@ -1,24 +1,22 @@
 package com.example.calculadorasimples
 
 class Calcular {
-        fun somar(N1 : Int, N2 : Int): Int{
+        fun somar(N1 : Float, N2 : Float): Float{
             val result =  N1 + N2
             return(result)
         }
 
-        fun multiplicar(N1 : Int, N2 : Int): Int
-        {
+        fun multiplicar(N1 : Float, N2 : Float): Float{
             val result = N1 * N2
             return(result)
         }
 
-        fun dividir(N1 : Int, N2 : Int): Int
-        {
+        fun dividir(N1 : Float, N2 : Float): Float{
             val result = N1 / N2
             return(result)
         }
 
-        fun subtrair(N1 : Int, N2 : Int): Int{
+        fun subtrair(N1 : Float, N2 : Float): Float{
             val result = N1 - N2
             return(result)
         }
